@@ -16,15 +16,17 @@
       "issues": "https://github.com/flextype-plugins/site/issues"
     },
     "require": {
-        "php": ">=7.2.0",
+        "php": ">=7.3.0",
         "ext-json": "*",
-        "middlewares/trailing-slash": "~1.1.0"
+        "middlewares/trailing-slash": "~1.1.0",
+        "flextype-components/arrays" : "3.0.1",
+        "flextype-components/filesystem": "2.0.8"
     },
     "config": {
         "apcu-autoloader": true,
         "optimize-autoloader": true,
         "platform": {
-            "php": "7.2.0"
+            "php": "7.3.0"
         }
     },
     "autoload": {
