@@ -16,16 +16,18 @@
       "issues": "https://github.com/flextype-plugins/twig/issues"
     },
     "require": {
-        "php": ">=7.2.0",
+        "php": ">=7.3.0",
         "slim/twig-view": "~2.5.0",
         "slim/csrf": "~0.8.3",
-        "slim/flash": "~0.4.0"
+        "slim/flash": "~0.4.0",
+        "flextype-components/arrays" : "3.0.1",
+        "flextype-components/filesystem": "2.0.8"
     },
     "config": {
         "apcu-autoloader": true,
         "optimize-autoloader": true,
         "platform": {
-            "php": "7.2.0"
+            "php": "7.3.0"
         }
     },
     "autoload": {
