@@ -5,6 +5,7 @@
     "keywords": ["templating"],
     "homepage": "https://twig.symfony.com",
     "license": "BSD-3-Clause",
+    "minimum-stability": "dev",
     "authors": [
         {
             "name": "Fabien Potencier",
@@ -23,18 +24,15 @@
         }
     ],
     "require": {
-        "php": "^7.0",
+        "php": ">=7.2.5",
         "symfony/polyfill-mbstring": "^1.3",
         "symfony/polyfill-ctype": "^1.8"
     },
     "require-dev": {
-        "symfony/phpunit-bridge": "^4.4|^5.0",
+        "symfony/phpunit-bridge": "^4.4.9|^5.0.9",
         "psr/container": "^1.0"
     },
     "autoload": {
-        "psr-0" : {
-            "Twig_" : "lib/"
-        },
         "psr-4" : {
             "Twig\\" : "src/"
         }
@@ -46,7 +44,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "2.12-dev"
+            "dev-master": "3.1-dev"
         }
     }
 }
