@@ -21,12 +21,13 @@
         }
     ],
     "require": {
-        "php": ">=5.5.9",
-        "ext-fileinfo": "*"
+        "php": "^7.2.5 || ^8.0",
+        "ext-fileinfo": "*",
+        "league/mime-type-detection": "^1.3"
     },
     "require-dev": {
-        "phpspec/phpspec": "^3.4",
-        "phpunit/phpunit": "^5.7.26"
+        "phpspec/prophecy": "^1.11.1",
+        "phpunit/phpunit": "^8.5.8"
     },
     "autoload": {
         "psr-4": {
@@ -36,10 +37,7 @@
     "autoload-dev": {
         "psr-4": {
             "League\\Flysystem\\Stub\\": "stub/"
-        },
-        "files": [
-          "tests/PHPUnitHacks.php"
-        ]
+        }
     },
     "suggest": {
         "ext-fileinfo": "Required for MimeType",
