@@ -13,7 +13,7 @@
     ],
     "homepage": "http://glide.thephpleague.com",
     "license": "MIT",
-    "authors" : [
+    "authors": [
         {
             "name": "Jonathan Reinink",
             "email": "jonathan@reinink.ca",
@@ -23,13 +23,13 @@
     "require": {
         "intervention/image": "^2.4",
         "league/flysystem": "^1.0",
-        "php": "^5.5 | ^7.0",
+        "php": "^7.2|^8.0",
         "psr/http-message": "^1.0"
     },
     "require-dev": {
-        "mockery/mockery": "~0.9",
-        "phpunit/php-token-stream": "^1.4",
-        "phpunit/phpunit": "~4.4"
+        "mockery/mockery": "^1.3.3",
+        "phpunit/phpunit": "^8.5|^9.0",
+        "phpunit/php-token-stream": "^3.1|^4.0"
     },
     "autoload": {
         "psr-4": {
