@@ -3,27 +3,42 @@
 	"description": "PHPDoc parser with support for nullable, intersection and generic types",
 	"license": "MIT",
 	"require": {
-		"php": "~7.1"
+		"php": "^7.1 || ^8.0"
 	},
 	"require-dev": {
 		"consistence/coding-standard": "^3.5",
+		"ergebnis/composer-normalize": "^2.0.2",
 		"jakub-onderka/php-parallel-lint": "^0.9.2",
 		"phing/phing": "^2.16.0",
-		"phpstan/phpstan": "^0.10",
+		"phpstan/extension-installer": "^1.0",
+		"phpstan/phpstan": "^0.12.26",
+		"phpstan/phpstan-strict-rules": "^0.12",
 		"phpunit/phpunit": "^6.3",
 		"slevomat/coding-standard": "^4.7.2",
-		"squizlabs/php_codesniffer": "^3.3.2",
-		"symfony/process": "^3.4 || ^4.0"
+		"symfony/process": "^4.0"
 	},
-	"autoload": {
-		"psr-4": {"PHPStan\\PhpDocParser\\": ["src/"]}
-	},
-	"autoload-dev": {
-		"psr-4": {"PHPStan\\PhpDocParser\\": ["tests/PHPStan"]}
+	"config": {
+		"sort-packages": true
 	},
 	"extra": {
 		"branch-alias": {
-			"dev-master": "0.3-dev"
+			"dev-master": "0.4-dev"
 		}
-	}
+	},
+	"autoload": {
+		"psr-4": {
+			"PHPStan\\PhpDocParser\\": [
+				"src/"
+			]
+		}
+	},
+	"autoload-dev": {
+		"psr-4": {
+			"PHPStan\\PhpDocParser\\": [
+				"tests/PHPStan"
+			]
+		}
+	},
+	"minimum-stability": "dev",
+	"prefer-stable": true
 }
