@@ -7,33 +7,28 @@ declare(strict_types=1);
  * Founded by Sergey Romanenko and maintained by Flextype Community.
  */
 
-namespace Flextype;
+namespace Flextype\Plugin\Twig\Twig;
 
-use Twig_Extension;
-use Twig_Extension_GlobalsInterface;
+use Twig\Extension\AbstractExtension;
+use Twig\Extension\GlobalsInterface;
 
-class CacheTwigExtension extends Twig_Extension implements Twig_Extension_GlobalsInterface
+class CacheTwigExtension extends AbstractExtension implements GlobalsInterface
 {
-    /**
-     * Flextype Dependency Container
-     */
-    private $flextype;
-
     /**
      * Constructor
      */
-    public function __construct($flextype)
+    public function __construct()
     {
-        $this->flextype = $flextype;
+
     }
 
     /**
      * Register Global variables in an extension
      */
-    public function getGlobals()
+    public function getGlobals() : array
     {
         return [
-            'cache' => $this->flextype['cache'],
+            'cache' => flextype('cache'),
         ];
     }
 }
