@@ -7,37 +7,31 @@ declare(strict_types=1);
  * Founded by Sergey Romanenko and maintained by Flextype Community.
  */
 
-namespace Flextype;
+namespace Flextype\Plugin\Twig\Twig;
 
-use Twig_Extension;
-use Twig_Extension_GlobalsInterface;
-use Twig_SimpleFunction;
+use Twig\Extension\AbstractExtension;
+use Twig\Extension\GlobalsInterface;
 
-class CsrfTwigExtension extends Twig_Extension implements Twig_Extension_GlobalsInterface
+class CsrfTwigExtension extends AbstractExtension implements GlobalsInterface
 {
-    /**
-     * Flextype Dependency Container
-     */
-    private $flextype;
-
     /**
      * Constructor
      */
-    public function __construct($flextype)
+    public function __construct()
     {
-        $this->flextype = $flextype;
+
     }
 
     /**
      * Register Global variables in an extension
      */
-    public function getGlobals()
+    public function getGlobals() : array
     {
         // CSRF token name and value
-        $csrfNameKey  = $this->flextype->csrf->getTokenNameKey();
-        $csrfValueKey = $this->flextype->csrf->getTokenValueKey();
-        $csrfName     = $this->flextype->csrf->getTokenName();
-        $csrfValue    = $this->flextype->csrf->getTokenValue();
+        $csrfNameKey  = flextype('csrf')->getTokenNameKey();
+        $csrfValueKey = flextype('csrf')->getTokenValueKey();
+        $csrfName     = flextype('csrf')->getTokenName();
+        $csrfValue    = flextype('csrf')->getTokenValue();
 
         return [
             'csrf'   => [
@@ -64,7 +58,7 @@ class CsrfTwigExtension extends Twig_Extension implements Twig_Extension_Globals
     public function getFunctions() : array
     {
         return [
-            new Twig_SimpleFunction('csrf', [$this, 'csrf'], ['is_safe' => ['html']]),
+            new \Twig\TwigFunction('csrf', [$this, 'csrf'], ['is_safe' => ['html']]),
         ];
     }
 
@@ -73,7 +67,7 @@ class CsrfTwigExtension extends Twig_Extension implements Twig_Extension_Globals
      */
     public function csrf() : string
     {
-        return '<input type="hidden" name="' . $this->flextype->csrf->getTokenNameKey() . '" value="' . $this->flextype->csrf->getTokenName() . '">' .
-               '<input type="hidden" name="' . $this->flextype->csrf->getTokenValueKey() . '" value="' . $this->flextype->csrf->getTokenValue() . '">';
+        return '<input type="hidden" name="' . flextype('csrf')->getTokenNameKey() . '" value="' . flextype('csrf')->getTokenName() . '">' .
+               '<input type="hidden" name="' . flextype('csrf')->getTokenValueKey() . '" value="' . flextype('csrf')->getTokenValue() . '">';
     }
 }
