@@ -7,24 +7,18 @@ declare(strict_types=1);
  * Founded by Sergey Romanenko and maintained by Flextype Community.
  */
 
-namespace Flextype;
+namespace Flextype\Plugin\Twig\Twig;
 
-use Twig_Extension;
-use Twig_SimpleFunction;
+use Twig\Extension\AbstractExtension;
 
-class FlashTwigExtension extends Twig_Extension
+class FlashTwigExtension extends AbstractExtension
 {
-    /**
-     * Flextype Dependency Container
-     */
-    private $flextype;
-
     /**
      * Constructor
      */
-    public function __construct($flextype)
+    public function __construct()
     {
-        $this->flextype = $flextype;
+
     }
 
     /**
@@ -35,7 +29,7 @@ class FlashTwigExtension extends Twig_Extension
     public function getFunctions() : array
     {
         return [
-            new Twig_SimpleFunction('flash', [$this, 'getMessages']),
+            new \Twig\TwigFunction('flash', [$this, 'getMessages']),
         ];
     }
 
@@ -48,9 +42,9 @@ class FlashTwigExtension extends Twig_Extension
     public function getMessages(?string $key = null) : array
     {
         if ($key !== null) {
-            return $this->flextype['flash']->getMessage($key);
+            return flextype('flash')->getMessage($key);
         }
 
-        return $this->flextype['flash']->getMessages();
+        return flextype('flash')->getMessages();
     }
 }
