@@ -7,37 +7,29 @@ declare(strict_types=1);
  * Founded by Sergey Romanenko and maintained by Flextype Community.
  */
 
-namespace Flextype;
+namespace Flextype\Plugin\Twig\Twig;
 
-use Twig_Extension;
-use Twig_Extension_GlobalsInterface;
+use Twig\Extension\AbstractExtension;
+use Twig\Extension\GlobalsInterface;
 use const PHP_VERSION;
 
-class GlobalVarsTwigExtension extends Twig_Extension implements Twig_Extension_GlobalsInterface
+class GlobalVarsTwigExtension extends AbstractExtension implements GlobalsInterface
 {
-    /**
-     * Flextype Dependency Container
-     */
-    private $flextype;
-
     /**
      * Constructor
      */
-    public function __construct($flextype)
+    public function __construct()
     {
-        $this->flextype = $flextype;
+
     }
 
     /**
      * Register Global variables in an extension
      */
-    public function getGlobals()
+    public function getGlobals() : array
     {
         return [
             'PATH_PROJECT' => PATH['project'],
-            'PATH_CONFIG' => PATH['config'],
-            'PATH_CACHE' => PATH['cache'],
-            'PATH_LOGS' => PATH['logs'],
             'PHP_VERSION' => PHP_VERSION
         ];
     }
