@@ -7,25 +7,18 @@ declare(strict_types=1);
  * Founded by Sergey Romanenko and maintained by Flextype Community.
  */
 
-namespace Flextype;
+namespace Flextype\Plugin\Twig\Twig;
 
-use Twig_Extension;
-use Twig_SimpleFilter;
-use Twig_SimpleFunction;
+use Twig\Extension\AbstractExtension;
 
-class JsonTwigExtension extends Twig_Extension
+class JsonTwigExtension extends AbstractExtension
 {
-    /**
-     * Flextype Dependency Container
-     */
-    private $flextype;
-
     /**
      * Constructor
      */
-    public function __construct($flextype)
+    public function __construct()
     {
-        $this->flextype = $flextype;
+
     }
 
     /**
@@ -36,8 +29,8 @@ class JsonTwigExtension extends Twig_Extension
     public function getFunctions() : array
     {
         return [
-            new Twig_SimpleFunction('json_decode', [$this, 'decode']),
-            new Twig_SimpleFunction('json_encode', [$this, 'encode']),
+            new \Twig\TwigFunction('json_decode', [$this, 'decode']),
+            new \Twig\TwigFunction('json_encode', [$this, 'encode']),
         ];
     }
 
@@ -49,8 +42,8 @@ class JsonTwigExtension extends Twig_Extension
     public function getFilters() : array
     {
         return [
-            new Twig_SimpleFilter('json_decode', [$this, 'decode']),
-            new Twig_SimpleFilter('json_encode', [$this, 'encode']),
+            new \Twig\TwigFilter('json_decode', [$this, 'decode']),
+            new \Twig\TwigFilter('json_encode', [$this, 'encode']),
         ];
     }
 
@@ -59,7 +52,7 @@ class JsonTwigExtension extends Twig_Extension
      */
     public function encode($input) : string
     {
-        return $this->flextype['serializer']->encode($input, 'json');
+        return flextype('json')->encode($input);
     }
 
     /**
@@ -67,6 +60,6 @@ class JsonTwigExtension extends Twig_Extension
      */
     public function decode(string $input, bool $cache = true)
     {
-        return $this->flextype['serializer']->decode($input, 'json', $cache);
+        return flextype('json')->decode($input, $cache);
     }
 }
