@@ -7,33 +7,29 @@ declare(strict_types=1);
  * Founded by Sergey Romanenko and maintained by Flextype Community.
  */
 
-namespace Flextype;
+namespace Flextype\Plugin\Twig\Twig;
 
-use Twig_Extension;
-use Twig_Extension_GlobalsInterface;
+use Twig\Extension\AbstractExtension;
+use Twig\Extension\GlobalsInterface;
 
-class RegistryTwigExtension extends Twig_Extension implements Twig_Extension_GlobalsInterface
+class RegistryTwigExtension extends AbstractExtension implements GlobalsInterface
 {
-    /**
-     * Flextype Dependency Container
-     */
-    private $flextype;
 
     /**
      * Constructor
      */
-    public function __construct($flextype)
+    public function __construct()
     {
-        $this->flextype = $flextype;
+
     }
 
     /**
      * Register Global variables in an extension
      */
-    public function getGlobals()
+    public function getGlobals() : array
     {
         return [
-            'registry' => new RegistryTwig($this->flextype),
+            'registry' => new RegistryTwig(),
         ];
     }
 }
@@ -41,16 +37,16 @@ class RegistryTwigExtension extends Twig_Extension implements Twig_Extension_Glo
 class RegistryTwig
 {
     /**
-     * Flextype Dependency Container
+     * Flextype Application
      */
-    private $flextype;
+
 
     /**
      * Constructor
      */
-    public function __construct($flextype)
+    public function __construct()
     {
-        $this->flextype = $flextype;
+
     }
 
     /**
@@ -60,7 +56,7 @@ class RegistryTwig
      */
     public function dump() : array
     {
-        return $this->flextype['registry']->dump();
+        return flextype('registry')->dump();
     }
 
     /**
@@ -71,7 +67,7 @@ class RegistryTwig
      */
     public function has(string $name) : bool
     {
-        return $this->flextype['registry']->has($name);
+        return flextype('registry')->has($name);
     }
 
     /**
@@ -83,7 +79,7 @@ class RegistryTwig
      */
     public function set(string $name, $value = null) : void
     {
-        $this->flextype['registry']->set($name, $value);
+        flextype('registry')->set($name, $value);
     }
 
     /**
@@ -95,6 +91,6 @@ class RegistryTwig
      */
     public function get(string $name, $default = null)
     {
-        return $this->flextype['registry']->get($name, $default);
+        return flextype('registry')->get($name, $default);
     }
 }
