@@ -7,26 +7,20 @@ declare(strict_types=1);
  * Founded by Sergey Romanenko and maintained by Flextype Community.
  */
 
-namespace Flextype;
+namespace Flextype\Plugin\Twig\Twig;
 
-use Twig_Extension;
-use Twig_SimpleFunction;
+use Twig\Extension\AbstractExtension;
 use Slim\Http\Environment;
 use Slim\Http\Uri;
 
-class UrlTwigExtension extends Twig_Extension
+class UrlTwigExtension extends AbstractExtension
 {
-    /**
-     * Flextype Dependency Container
-     */
-    private $flextype;
-
     /**
      * Constructor
      */
-    public function __construct($flextype)
+    public function __construct()
     {
-        $this->flextype = $flextype;
+
     }
 
     /**
@@ -37,7 +31,7 @@ class UrlTwigExtension extends Twig_Extension
     public function getFunctions() : array
     {
         return [
-            new Twig_SimpleFunction('url', [$this, 'url'], ['is_safe' => ['html']])
+            new \Twig\TwigFunction('url', [$this, 'url'], ['is_safe' => ['html']])
         ];
     }
 
@@ -46,8 +40,8 @@ class UrlTwigExtension extends Twig_Extension
      */
     public function url() : string
     {
-        if ($this->flextype->registry->has('flextype.settings.url') && $this->flextype->registry->get('flextype.settings.url') != '') {
-            return $this->flextype->registry->get('flextype.settings.url');
+        if (flextype('registry')->has('flextype.settings.url') && flextype('registry')->get('flextype.settings.url') != '') {
+            return flextype('registry')->get('flextype.settings.url');
         } else {
             return Uri::createFromEnvironment(new Environment($_SERVER))->getBaseUrl();
         }
