@@ -7,25 +7,18 @@ declare(strict_types=1);
  * Founded by Sergey Romanenko and maintained by Flextype Community.
  */
 
-namespace Flextype;
+namespace Flextype\Plugin\Twig\Twig;
 
-use Twig_Extension;
-use Twig_SimpleFilter;
-use Twig_SimpleFunction;
+use Twig\Extension\AbstractExtension;
 
-class YamlTwigExtension extends Twig_Extension
+class YamlTwigExtension extends AbstractExtension
 {
-    /**
-     * Flextype Dependency Container
-     */
-    private $flextype;
-
     /**
      * Constructor
      */
-    public function __construct($flextype)
+    public function __construct()
     {
-        $this->flextype = $flextype;
+
     }
 
     /**
@@ -36,8 +29,8 @@ class YamlTwigExtension extends Twig_Extension
     public function getFunctions() : array
     {
         return [
-            new Twig_SimpleFunction('yaml_decode', [$this, 'decode']),
-            new Twig_SimpleFunction('yaml_encode', [$this, 'encode']),
+            new \Twig\TwigFunction('yaml_decode', [$this, 'decode']),
+            new \Twig\TwigFunction('yaml_encode', [$this, 'encode']),
         ];
     }
 
@@ -49,8 +42,8 @@ class YamlTwigExtension extends Twig_Extension
     public function getFilters() : array
     {
         return [
-            new Twig_SimpleFilter('yaml_decode', [$this, 'decode']),
-            new Twig_SimpleFilter('yaml_encode', [$this, 'encode']),
+            new \Twig\TwigFilter('yaml_decode', [$this, 'decode']),
+            new \Twig\TwigFilter('yaml_encode', [$this, 'encode']),
         ];
     }
 
@@ -59,7 +52,7 @@ class YamlTwigExtension extends Twig_Extension
      */
     public function encode($input) : string
     {
-        return $this->flextype['serializer']->encode($input, 'yaml');
+        return flextype('yaml')->encode($input);
     }
 
     /**
@@ -67,6 +60,6 @@ class YamlTwigExtension extends Twig_Extension
      */
     public function decode(string $input, bool $cache = true)
     {
-        return $this->flextype['serializer']->decode($input, 'yaml', $cache);
+        return flextype('yaml')->decode($input, $cache);
     }
 }
