@@ -9,12 +9,21 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Flextype;
+namespace Flextype\Plugin\Icon\Models;
 
 use Flextype\Component\Filesystem\Filesystem;
 
-class Icon extends Container
+class Icon
 {
+
+    /**
+     * __construct
+     */
+     public function __construct()
+     {
+
+     }
+
     public static function icon($value = null) : string
     {
         $icon_fallback_file_path = PATH['project'] . '/plugins/icon/assets/dist/fontawesome/svgs/regular/file-alt.svg';
