@@ -2,4 +2,4 @@
 
 namespace Flextype;
 
-$app->get('/' . $flextype->registry->get('plugins.sitemap.settings.route'), 'SitemapController:index')->setName('sitemap.index');
+flextype()->get('/' . flextype('registry')->get('plugins.sitemap.settings.route'), 'SitemapController:index')->setName('sitemap.index');
