@@ -9,17 +9,20 @@ declare(strict_types=1);
 
 namespace Flextype;
 
-use Flextype\Component\Filesystem\Filesystem;
-use Psr\Http\Message\ResponseInterface as Response;
 use Psr\Http\Message\ServerRequestInterface as Request;
+use Slim\Http\Response;
+
 use function array_replace_recursive;
+use function count;
+use function filesystem;
+use function flextype;
 
 /**
  * Validate folders token
  */
-function validate_folders_token($token) : bool
+function validate_folders_token($token): bool
 {
-    return Filesystem::has(PATH['project'] . '/tokens/folders/' . $token . '/token.yaml');
+    return filesystem()->file(PATH['project'] . '/tokens/folders/' . $token . '/token.yaml')->exists();
 }
 
 /**
@@ -28,77 +31,104 @@ function validate_folders_token($token) : bool
  * endpoint: GET /api/folders
  *
  * Query:
- * path   - [REQUIRED] - Folder path.
- * mode   - [REQUIRED] - Mode.
- * token  - [REQUIRED] - Valid Files token.
+ * path       - [REQUIRED] - Folder path.
+ * collection - [OPTIONAL] - Collection or single.
+ * token      - [REQUIRED] - Valid Folders token.
  *
  * Returns:
  * An array of folder(s) item objects.
  */
-$app->get('/api/folders', function (Request $request, Response $response) use ($flextype, $api_sys_messages) {
-
+flextype()->get('/api/folders', function (Request $request, Response $response) use ($api_errors) {
     // Get Query Params
     $query = $request->getQueryParams();
 
+    if (! isset($query['path']) || ! isset($query['token'])) {
+        return $response->withStatus($api_errors['0600']['http_status_code'])
+                        ->withHeader('Content-Type', 'application/json;charset=' . flextype('registry')->get('flextype.settings.charset'))
+                        ->write(flextype('json')->encode($api_errors['0600']));
+    }
+
     // Set variables
     $path  = $query['path'];
-    $mode  = $query['mode'];
     $token = $query['token'];
 
-    if ($flextype['registry']->get('flextype.settings.api.folders.enabled')) {
+    if (isset($query['collection'])) {
+        if ($query['collection'] === 'true') {
+            $collection = true;
+        } else {
+            $collection = false;
+        }
+    } else {
+        $collection = false;
+    }
 
+    if (flextype('registry')->get('flextype.settings.api.folders.enabled')) {
         // Validate delivery token
         if (validate_folders_token($token)) {
-            $folders_token_file_path = PATH['project'] . '/tokens/folders/' . $token. '/token.yaml';
+            $folders_token_file_path = PATH['project'] . '/tokens/folders/' . $token . '/token.yaml';
 
             // Set delivery token file
-            if ($folders_token_file_data = $flextype['serializer']->decode(Filesystem::read($folders_token_file_path), 'yaml')) {
-                if ($folders_token_file_data['state'] === 'disabled' ||
-                    ($folders_token_file_data['limit_calls'] !== 0 && $folders_token_file_data['calls'] >= $folders_token_file_data['limit_calls'])) {
-                    return $response->withJson($api_sys_messages['AccessTokenInvalid'], 401);
+            if ($folders_token_file_data = flextype('yaml')->decode(filesystem()->file($folders_token_file_path)->get())) {
+                if (
+                    $folders_token_file_data['state'] === 'disabled' ||
+                    ($folders_token_file_data['limit_calls'] !== 0 && $folders_token_file_data['calls'] >= $folders_token_file_data['limit_calls'])
+                ) {
+                    return $response->withStatus($api_errors['0501']['http_status_code'])
+                                    ->withHeader('Content-Type', 'application/json;charset=' . flextype('registry')->get('flextype.settings.charset'))
+                                    ->write(flextype('json')->encode($api_errors['0501']));
                 }
 
                 // Create folders array
                 $folders = [];
 
                 // Get list if folder or fodlers for specific folder
-                if ($mode == 'collection') {
-                    $folders = $flextype['media_folders']->fetchCollection($path);
-                } elseif ($mode == 'single') {
-                    $folders = $flextype['media_folders']->fetchSingle($path);
+                if ($collection) {
+                    $folders = flextype('media_folders')->fetchCollection($path)->toArray();
+                } else {
+                    $folders = flextype('media_folders')->fetchSingle($path)->toArray();
                 }
 
                 // Write response data
+                $response_data         = [];
                 $response_data['data'] = $folders;
 
                 // Set response code
                 $response_code = count($response_data['data']) > 0 ? 200 : 404;
 
                 // Update calls counter
-                Filesystem::write($folders_token_file_path, $flextype['serializer']->encode(array_replace_recursive($folders_token_file_data, ['calls' => $folders_token_file_data['calls'] + 1]), 'yaml'));
+                filesystem()->file($folders_token_file_path)->put(flextype('yaml')->encode(array_replace_recursive($folders_token_file_data, ['calls' => $folders_token_file_data['calls'] + 1])));
 
-                if ($response_code == 404) {
-
+                if ($response_code === 404) {
                     // Return response
                     return $response
-                           ->withJson($api_sys_messages['NotFound'], $response_code);
+                           ->withStatus($api_errors['0602']['http_status_code'])
+                           ->withHeader('Content-Type', 'application/json;charset=' . flextype('registry')->get('flextype.settings.charset'))
+                           ->write(flextype('json')->encode($api_errors['0602']));
                 }
 
                 // Return response
                 return $response
-                       ->withJson($response_data, $response_code);
+                       ->withStatus($response_code)
+                       ->withHeader('Content-Type', 'application/json;charset=' . flextype('registry')->get('flextype.settings.charset'))
+                       ->write(flextype('json')->encode($response_data));
             }
 
             return $response
-                   ->withJson($api_sys_messages['AccessTokenInvalid'], 401);
+                   ->withStatus($api_errors['0003']['http_status_code'])
+            ->withHeader('Content-Type', 'application/json;charset=' . flextype('registry')->get('flextype.settings.charset'))
+            ->write(flextype('json')->encode($api_errors['0003']));
         }
 
         return $response
-               ->withJson($api_sys_messages['AccessTokenInvalid'], 401);
+               ->withStatus($api_errors['0003']['http_status_code'])
+            ->withHeader('Content-Type', 'application/json;charset=' . flextype('registry')->get('flextype.settings.charset'))
+            ->write(flextype('json')->encode($api_errors['0003']));
     }
 
     return $response
-           ->withJson($api_sys_messages['AccessTokenInvalid'], 401);
+           ->withStatus($api_errors['0003']['http_status_code'])
+            ->withHeader('Content-Type', 'application/json;charset=' . flextype('registry')->get('flextype.settings.charset'))
+            ->write(flextype('json')->encode($api_errors['0003']));
 });
 
 
@@ -109,248 +139,415 @@ $app->get('/api/folders', function (Request $request, Response $response) use ($
  *
  * Body:
  * path          - [REQUIRED] - New folder path.
- * token         - [REQUIRED] - Valid Entries token.
+ * token         - [REQUIRED] - Valid Folders token.
  * access_token  - [REQUIRED] - Valid Access token.
  *
  * Returns:
- * Returns the file object for the file that was just created.
+ * Returns the folder object for the folder that was just created.
  */
-$app->post('/api/folders', function (Request $request, Response $response) use ($flextype, $api_sys_messages) {
+flextype()->post('/api/folders', function (Request $request, Response $response) use ($api_errors) {
+    // Get Post Data
+    $post_data = $request->getParsedBody();
 
-  // Get Post Data
-  $post_data = $request->getParsedBody();
+    if (! isset($post_data['token']) || ! isset($post_data['access_token']) || ! isset($post_data['path'])) {
+        return $response->withStatus($api_errors['0601']['http_status_code'])
+            ->withHeader('Content-Type', 'application/json;charset=' . flextype('registry')->get('flextype.settings.charset'))
+            ->write(flextype('json')->encode($api_errors['0601']));
+    }
 
-  // Set variables
-  $token        = $post_data['token'];
-  $access_token = $post_data['access_token'];
-  $path         = $post_data['path'];
+    // Set variables
+    $token        = $post_data['token'];
+    $access_token = $post_data['access_token'];
+    $path         = $post_data['path'];
 
-  if ($flextype['registry']->get('flextype.settings.api.files.enabled')) {
+    if (flextype('registry')->get('flextype.settings.api.folders.enabled')) {
+        // Validate files and access token
+        if (validate_folders_token($token) && validate_access_token($access_token)) {
+            $folders_token_file_path = PATH['project'] . '/tokens/folders/' . $token . '/token.yaml';
+            $access_token_file_path  = PATH['project'] . '/tokens/access/' . $access_token . '/token.yaml';
 
-      // Validate files and access token
-      if (validate_files_token($token) && validate_access_token($access_token)) {
-          $files_token_file_path = PATH['project'] . '/tokens/files/' . $token . '/token.yaml';
-          $access_token_file_path = PATH['project'] . '/tokens/access/' . $access_token . '/token.yaml';
+            // Set files and access token file
+            if (
+                ($folders_token_file_data = flextype('yaml')->decode(filesystem()->file($folders_token_file_path)->get())) &&
+                ($access_token_file_data = flextype('yaml')->decode(filesystem()->file($access_token_file_path)->get()))
+            ) {
+                if (
+                    $folders_token_file_data['state'] === 'disabled' ||
+                    ($folders_token_file_data['limit_calls'] !== 0 && $folders_token_file_data['calls'] >= $folders_token_file_data['limit_calls'])
+                ) {
+                    return $response->withStatus($api_errors['0501']['http_status_code'])
+                                    ->withHeader('Content-Type', 'application/json;charset=' . flextype('registry')->get('flextype.settings.charset'))
+                                    ->write(flextype('json')->encode($api_errors['0501']));
+                }
 
-          // Set files and access token file
-          if (($files_token_file_data = $flextype['serializer']->decode(Filesystem::read($files_token_file_path), 'yaml')) &&
-              ($access_token_file_data = $flextype['serializer']->decode(Filesystem::read($access_token_file_path), 'yaml'))) {
+                if (
+                    $access_token_file_data['state'] === 'disabled' ||
+                    ($access_token_file_data['limit_calls'] !== 0 && $access_token_file_data['calls'] >= $access_token_file_data['limit_calls'])
+                ) {
+                    return $response->withStatus($api_errors['0501']['http_status_code'])
+                                    ->withHeader('Content-Type', 'application/json;charset=' . flextype('registry')->get('flextype.settings.charset'))
+                                    ->write(flextype('json')->encode($api_errors['0501']));
+                }
 
-              if ($files_token_file_data['state'] === 'disabled' ||
-                  ($files_token_file_data['limit_calls'] !== 0 && $files_token_file_data['calls'] >= $files_token_file_data['limit_calls'])) {
-                  return $response->withJson($api_sys_messages['AccessTokenInvalid'], 401);
-              }
+                // Create folder
+                $create_folder = flextype('media_folders')->create($path);
 
-              if ($access_token_file_data['state'] === 'disabled' ||
-                  ($access_token_file_data['limit_calls'] !== 0 && $access_token_file_data['calls'] >= $access_token_file_data['limit_calls'])) {
-                  return $response->withJson($api_sys_messages['AccessTokenInvalid'], 401);
-              }
+                $response_data = [];
 
-              // Create folder
-              $create_folder = $flextype['media_folders']->create($path);
+                if ($create_folder) {
+                    $response_data['data'] = flextype('media_folders')->fetchSingle($path);
+                }
 
-              if ($create_folder) {
-                  $response_data['data'] = $flextype['media_folders']->fetch($path);
-              } else {
-                  $response_data['data'] = [];
-              }
+                // Set response code
+                $response_code = $create_folder ? 200 : 404;
 
-              // Set response code
-              $response_code = ($create_folder) ? 200 : 404;
+                // Update calls counter
+                filesystem()->file($folders_token_file_path)->put(flextype('yaml')->encode(array_replace_recursive($folders_token_file_data, ['calls' => $folders_token_file_data['calls'] + 1])));
 
-              // Return response
-              return $response
-                     ->withJson($response_data, $response_code);
+                if ($response_code === 404) {
+                    // Return response
+                    return $response
+                         ->withStatus($api_errors['0602']['http_status_code'])
+                        ->withHeader('Content-Type', 'application/json;charset=' . flextype('registry')->get('flextype.settings.charset'))
+                        ->write(flextype('json')->encode($api_errors['0602']));
+                }
 
-              // Update calls counter
-              Filesystem::write($files_token_file_path, $flextype['serializer']->encode(array_replace_recursive($files_token_file_data, ['calls' => $files_token_file_data['calls'] + 1]), 'yaml'));
+                // Return response
+                return $response
+                     ->withStatus($response_code)
+                     ->withHeader('Content-Type', 'application/json;charset=' . flextype('registry')->get('flextype.settings.charset'))
+                     ->write(flextype('json')->encode($response_data));
+            }
 
-              if ($response_code == 404) {
+            return $response
+                        ->withStatus($api_errors['0003']['http_status_code'])
+                        ->withHeader('Content-Type', 'application/json;charset=' . flextype('registry')->get('flextype.settings.charset'))
+                        ->write(flextype('json')->encode($api_errors['0003']));
+        }
 
-                  // Return response
-                  return $response
-                         ->withJson($api_sys_messages['NotFound'], $response_code);
-              }
+        return $response
+                    ->withStatus($api_errors['0003']['http_status_code'])
+                    ->withHeader('Content-Type', 'application/json;charset=' . flextype('registry')->get('flextype.settings.charset'))
+                    ->write(flextype('json')->encode($api_errors['0003']));
+    }
 
-              // Return response
-              return $response
-                     ->withJson($response_data, $response_code);
-          }
-
-          return $response
-                 ->withJson($api_sys_messages['AccessTokenInvalid'], 401);
-      }
-
-      return $response
-             ->withJson($api_sys_messages['AccessTokenInvalid'], 401);
-  }
-
-  return $response
-         ->withJson($api_sys_messages['AccessTokenInvalid'], 401);
+    return $response
+                ->withStatus($api_errors['0003']['http_status_code'])
+                ->withHeader('Content-Type', 'application/json;charset=' . flextype('registry')->get('flextype.settings.charset'))
+                ->write(flextype('json')->encode($api_errors['0003']));
 });
 
 /**
- * Rename folder
+ * Copy folder
+ *
+ * endpoint: PUT /api/folders/copy
+ *
+ * Body:
+ * path            - [REQUIRED] - Unique identifier of the folder.
+ * new_path        - [REQUIRED] - New Unique identifier of the folder.
+ * token           - [REQUIRED] - Valid Folder token.
+ * access_token    - [REQUIRED] - Valid Access token.
+ *
+ * Returns:
+ * Returns the folders object for the folders that was just copied.
+ */
+flextype()->put('/api/folders/copy', function (Request $request, Response $response) use ($api_errors) {
+    // Get Post Data
+    $post_data = $request->getParsedBody();
+
+    if (! isset($post_data['token']) || ! isset($post_data['access_token']) || ! isset($post_data['path']) || ! isset($post_data['new_path'])) {
+        return $response->withStatus($api_errors['0601']['http_status_code'])
+            ->withHeader('Content-Type', 'application/json;charset=' . flextype('registry')->get('flextype.settings.charset'))
+            ->write(flextype('json')->encode($api_errors['0601']));
+    }
+
+    // Set variables
+    $token        = $post_data['token'];
+    $access_token = $post_data['access_token'];
+    $path         = $post_data['path'];
+    $new_path     = $post_data['new_path'];
+
+    if (flextype('registry')->get('flextype.settings.api.folders.enabled')) {
+        // Validate files and access token
+        if (validate_folders_token($token) && validate_access_token($access_token)) {
+            $folders_token_file_path = PATH['project'] . '/tokens/folders/' . $token . '/token.yaml';
+            $access_token_file_path  = PATH['project'] . '/tokens/access/' . $access_token . '/token.yaml';
+
+            // Set files and access token file
+            if (
+                ($folders_token_file_data = flextype('yaml')->decode(filesystem()->file($folders_token_file_path)->get())) &&
+                ($access_token_file_data = flextype('yaml')->decode(filesystem()->file($access_token_file_path)->get()))
+            ) {
+                if (
+                    $folders_token_file_data['state'] === 'disabled' ||
+                    ($folders_token_file_data['limit_calls'] !== 0 && $folders_token_file_data['calls'] >= $folders_token_file_data['limit_calls'])
+                ) {
+                    return $response->withStatus($api_errors['0601']['http_status_code'])
+                    ->withHeader('Content-Type', 'application/json;charset=' . flextype('registry')->get('flextype.settings.charset'))
+                    ->write(flextype('json')->encode($api_errors['0601']));
+                }
+
+                if (
+                    $access_token_file_data['state'] === 'disabled' ||
+                    ($access_token_file_data['limit_calls'] !== 0 && $access_token_file_data['calls'] >= $access_token_file_data['limit_calls'])
+                ) {
+                    return $response->withStatus($api_errors['0601']['http_status_code'])
+                    ->withHeader('Content-Type', 'application/json;charset=' . flextype('registry')->get('flextype.settings.charset'))
+                    ->write(flextype('json')->encode($api_errors['0601']));
+                }
+
+                // Copy folder
+                $copy_folder = flextype('media_folders')->copy($path, $new_path);
+
+                $response_data = [];
+
+                if ($copy_folder) {
+                    $response_data['data'] = flextype('media_folders')->fetchSingle($new_path);
+                } else {
+                    $response_data['data'] = $copy_folder;
+                }
+
+                // Set response code
+                $response_code = $copy_folder ? 200 : 404;
+
+                // Update calls counter
+                filesystem()->file($folders_token_file_path)->put(flextype('yaml')->encode(array_replace_recursive($folders_token_file_data, ['calls' => $folders_token_file_data['calls'] + 1])));
+
+                if ($response_code === 404) {
+                    // Return response
+                    return $response
+                                ->withStatus($api_errors['0602']['http_status_code'])
+                                ->withHeader('Content-Type', 'application/json;charset=' . flextype('registry')->get('flextype.settings.charset'))
+                                ->write(flextype('json')->encode($api_errors['0602']));
+                }
+
+                // Return response
+                return $response
+                     ->withStatus($response_code)
+                ->withHeader('Content-Type', 'application/json;charset=' . flextype('registry')->get('flextype.settings.charset'))
+                ->write(flextype('json')->encode($response_data));
+            }
+
+            return $response
+                        ->withStatus($api_errors['0003']['http_status_code'])
+                        ->withHeader('Content-Type', 'application/json;charset=' . flextype('registry')->get('flextype.settings.charset'))
+                        ->write(flextype('json')->encode($api_errors['0003']));
+        }
+
+        return $response
+                    ->withStatus($api_errors['0003']['http_status_code'])
+                    ->withHeader('Content-Type', 'application/json;charset=' . flextype('registry')->get('flextype.settings.charset'))
+                    ->write(flextype('json')->encode($api_errors['0003']));
+    }
+
+    return $response
+                ->withStatus($api_errors['0003']['http_status_code'])
+                ->withHeader('Content-Type', 'application/json;charset=' . flextype('registry')->get('flextype.settings.charset'))
+                ->write(flextype('json')->encode($api_errors['0003']));
+});
+
+/**
+ * Move folder
  *
  * endpoint: PUT /api/folders
  *
  * Body:
- * id            - [REQUIRED] - Unique identifier of the file.
- * new_id        - [REQUIRED] - New Unique identifier of the file.
- * token         - [REQUIRED] - Valid Entries token.
- * access_token  - [REQUIRED] - Valid Access token.
+ * path            - [REQUIRED] - Unique identifier of the folder.
+ * new_pah         - [REQUIRED] - New Unique identifier of the folder.
+ * token           - [REQUIRED] - Valid Folder token.
+ * access_token    - [REQUIRED] - Valid Access token.
  *
  * Returns:
- * Returns the file object for the file that was just created.
+ * Returns the folders object for the folders that was just renamed.
  */
-$app->put('/api/folders', function (Request $request, Response $response) use ($flextype, $api_sys_messages) {
+flextype()->put('/api/folders', function (Request $request, Response $response) use ($api_errors) {
+    // Get Post Data
+    $post_data = $request->getParsedBody();
 
-  // Get Post Data
-  $post_data = $request->getParsedBody();
+    if (! isset($post_data['token']) || ! isset($post_data['access_token']) || ! isset($post_data['path']) || ! isset($post_data['new_path'])) {
+        return $response->withStatus($api_errors['0601']['http_status_code'])
+            ->withHeader('Content-Type', 'application/json;charset=' . flextype('registry')->get('flextype.settings.charset'))
+            ->write(flextype('json')->encode($api_errors['0601']));
+    }
 
-  // Set variables
-  $token        = $post_data['token'];
-  $access_token = $post_data['access_token'];
-  $id           = $post_data['id'];
-  $new_id       = $post_data['new_id'];
+    // Set variables
+    $token        = $post_data['token'];
+    $access_token = $post_data['access_token'];
+    $path         = $post_data['path'];
+    $new_path     = $post_data['new_path'];
 
-  if ($flextype['registry']->get('flextype.settings.api.files.enabled')) {
+    if (flextype('registry')->get('flextype.settings.api.folders.enabled')) {
+        // Validate files and access token
+        if (validate_folders_token($token) && validate_access_token($access_token)) {
+            $folders_token_file_path = PATH['project'] . '/tokens/folders/' . $token . '/token.yaml';
+            $access_token_file_path  = PATH['project'] . '/tokens/access/' . $access_token . '/token.yaml';
 
-      // Validate files and access token
-      if (validate_files_token($token) && validate_access_token($access_token)) {
-          $files_token_file_path = PATH['project'] . '/tokens/files/' . $token . '/token.yaml';
-          $access_token_file_path = PATH['project'] . '/tokens/access/' . $access_token . '/token.yaml';
+            // Set files and access token file
+            if (
+                ($folders_token_file_data = flextype('yaml')->decode(filesystem()->file($folders_token_file_path)->get())) &&
+                ($access_token_file_data = flextype('yaml')->decode(filesystem()->file($access_token_file_path)->get()))
+            ) {
+                if (
+                    $folders_token_file_data['state'] === 'disabled' ||
+                    ($folders_token_file_data['limit_calls'] !== 0 && $folders_token_file_data['calls'] >= $folders_token_file_data['limit_calls'])
+                ) {
+                    return $response->withStatus($api_errors['0501']['http_status_code'])
+                                    ->withHeader('Content-Type', 'application/json;charset=' . flextype('registry')->get('flextype.settings.charset'))
+                                    ->write(flextype('json')->encode($api_errors['0501']));
+                }
 
-          // Set files and access token file
-          if (($files_token_file_data = $flextype['serializer']->decode(Filesystem::read($files_token_file_path), 'yaml')) &&
-              ($access_token_file_data = $flextype['serializer']->decode(Filesystem::read($access_token_file_path), 'yaml'))) {
+                if (
+                    $access_token_file_data['state'] === 'disabled' ||
+                    ($access_token_file_data['limit_calls'] !== 0 && $access_token_file_data['calls'] >= $access_token_file_data['limit_calls'])
+                ) {
+                    return $response->withStatus($api_errors['0501']['http_status_code'])
+                                    ->withHeader('Content-Type', 'application/json;charset=' . flextype('registry')->get('flextype.settings.charset'))
+                                    ->write(flextype('json')->encode($api_errors['0501']));
+                }
 
-              if ($files_token_file_data['state'] === 'disabled' ||
-                  ($files_token_file_data['limit_calls'] !== 0 && $files_token_file_data['calls'] >= $files_token_file_data['limit_calls'])) {
-                  return $response->withJson($api_sys_messages['AccessTokenInvalid'], 401);
-              }
+                // Rename folder
+                $move_folder = flextype('media_folders')->move($path, $new_path);
 
-              if ($access_token_file_data['state'] === 'disabled' ||
-                  ($access_token_file_data['limit_calls'] !== 0 && $access_token_file_data['calls'] >= $access_token_file_data['limit_calls'])) {
-                  return $response->withJson($api_sys_messages['AccessTokenInvalid'], 401);
-              }
+                $response_data = [];
 
-              // Rename folder
-              $rename_folder = $flextype['media_folders']->rename($id, $new_id);
+                if ($move_folder) {
+                    $response_data['data'] = flextype('media_folders')->fetchSingle($new_path);
+                }
 
-              if ($rename_folder) {
-                  $response_data['data'] = $flextype['media_folders']->fetch($new_id);
-              } else {
-                  $response_data['data'] = [];
-              }
+                // Set response code
+                $response_code = $move_folder ? 200 : 404;
 
-              // Set response code
-              $response_code = ($rename_folder) ? 200 : 404;
+                // Update calls counter
+                filesystem()->file($folders_token_file_path)->put(flextype('yaml')->encode(array_replace_recursive($folders_token_file_data, ['calls' => $folders_token_file_data['calls'] + 1])));
 
-              // Return response
-              return $response
-                     ->withJson($response_data, $response_code);
+                if ($response_code === 404) {
+                    // Return response
+                    return $response
+                                ->withStatus($api_errors['0602']['http_status_code'])
+                                ->withHeader('Content-Type', 'application/json;charset=' . flextype('registry')->get('flextype.settings.charset'))
+                                ->write(flextype('json')->encode($api_errors['0602']));
+                }
 
-              // Update calls counter
-              Filesystem::write($files_token_file_path, $flextype['serializer']->encode(array_replace_recursive($files_token_file_data, ['calls' => $files_token_file_data['calls'] + 1]), 'yaml'));
+                // Return response
+                return $response
+                     ->withStatus($response_code)
+                ->withHeader('Content-Type', 'application/json;charset=' . flextype('registry')->get('flextype.settings.charset'))
+                ->write(flextype('json')->encode($response_data));
+            }
 
-              if ($response_code == 404) {
+            return $response
+                        ->withStatus($api_errors['0003']['http_status_code'])
+                        ->withHeader('Content-Type', 'application/json;charset=' . flextype('registry')->get('flextype.settings.charset'))
+                        ->write(flextype('json')->encode($api_errors['0003']));
+        }
 
-                  // Return response
-                  return $response
-                         ->withJson($api_sys_messages['NotFound'], $response_code);
-              }
+        return $response
+                    ->withStatus($api_errors['0003']['http_status_code'])
+                    ->withHeader('Content-Type', 'application/json;charset=' . flextype('registry')->get('flextype.settings.charset'))
+                    ->write(flextype('json')->encode($api_errors['0003']));
+    }
 
-              // Return response
-              return $response
-                     ->withJson($response_data, $response_code);
-          }
-
-          return $response
-                 ->withJson($api_sys_messages['AccessTokenInvalid'], 401);
-      }
-
-      return $response
-             ->withJson($api_sys_messages['AccessTokenInvalid'], 401);
-  }
-
-  return $response
-         ->withJson($api_sys_messages['AccessTokenInvalid'], 401);
+    return $response
+                ->withStatus($api_errors['0003']['http_status_code'])
+                ->withHeader('Content-Type', 'application/json;charset=' . flextype('registry')->get('flextype.settings.charset'))
+                ->write(flextype('json')->encode($api_errors['0003']));
 });
 
 /**
-* Delete file
+* Delete folder
 *
-* endpoint: DELETE /api/files
+* endpoint: DELETE /api/folders
 *
 * Body:
-* id           - [REQUIRED] - Unique identifier of the file.
-* token        - [REQUIRED] - Valid Entries token.
+* id           - [REQUIRED] - Unique identifier of the folder.
+* token        - [REQUIRED] - Valid Folders token.
 * access_token - [REQUIRED] - Valid Authentication token.
 *
 * Returns:
 * Returns an empty body with HTTP status 204
 */
-$app->delete('/api/folders', function (Request $request, Response $response) use ($flextype) {
+flextype()->delete('/api/folders', function (Request $request, Response $response) use ($api_errors) {
+    // Get Post Data
+    $post_data = $request->getParsedBody();
 
-  // Get Post Data
-  $post_data = $request->getParsedBody();
+    if (! isset($post_data['token']) || ! isset($post_data['access_token']) || ! isset($post_data['path'])) {
+        return $response->withStatus($api_errors['0601']['http_status_code'])
+            ->withHeader('Content-Type', 'application/json;charset=' . flextype('registry')->get('flextype.settings.charset'))
+            ->write(flextype('json')->encode($api_errors['0601']));
+    }
 
-  // Set variables
-  $token        = $post_data['token'];
-  $access_token = $post_data['access_token'];
-  $id           = $post_data['id'];
+    // Set variables
+    $token        = $post_data['token'];
+    $access_token = $post_data['access_token'];
+    $path         = $post_data['path'];
 
-  if ($flextype['registry']->get('flextype.settings.api.files.enabled')) {
+    if (flextype('registry')->get('flextype.settings.api.folders.enabled')) {
+        // Validate files and access token
+        if (validate_folders_token($token) && validate_access_token($access_token)) {
+            $folders_token_file_path = PATH['project'] . '/tokens/folders/' . $token . '/token.yaml';
+            $access_token_file_path  = PATH['project'] . '/tokens/access/' . $access_token . '/token.yaml';
 
-      // Validate files and access token
-      if (validate_files_token($token) && validate_access_token($access_token)) {
-          $files_token_file_path = PATH['project'] . '/tokens/files/' . $token . '/token.yaml';
-          $access_token_file_path = PATH['project'] . '/tokens/access/' . $access_token . '/token.yaml';
+            // Set files and access token file
+            if (
+                ($folders_token_file_data = flextype('yaml')->decode(filesystem()->file($folders_token_file_path)->get())) &&
+                ($access_token_file_data = flextype('yaml')->decode(filesystem()->file($access_token_file_path)->get()))
+            ) {
+                if (
+                    $folders_token_file_data['state'] === 'disabled' ||
+                    ($folders_token_file_data['limit_calls'] !== 0 && $folders_token_file_data['calls'] >= $folders_token_file_data['limit_calls'])
+                ) {
+                    return $response->withStatus($api_errors['0501']['http_status_code'])
+                                    ->withHeader('Content-Type', 'application/json;charset=' . flextype('registry')->get('flextype.settings.charset'))
+                                    ->write(flextype('json')->encode($api_errors['0501']));
+                }
 
-          // Set files and access token file
-          if (($files_token_file_data = $flextype['serializer']->decode(Filesystem::read($files_token_file_path), 'yaml')) &&
-              ($access_token_file_data = $flextype['serializer']->decode(Filesystem::read($access_token_file_path), 'yaml'))) {
+                if (
+                    $access_token_file_data['state'] === 'disabled' ||
+                    ($access_token_file_data['limit_calls'] !== 0 && $access_token_file_data['calls'] >= $access_token_file_data['limit_calls'])
+                ) {
+                    return $response->withStatus($api_errors['0501']['http_status_code'])
+                                    ->withHeader('Content-Type', 'application/json;charset=' . flextype('registry')->get('flextype.settings.charset'))
+                                    ->write(flextype('json')->encode($api_errors['0501']));
+                }
 
-              if ($files_token_file_data['state'] === 'disabled' ||
-                  ($files_token_file_data['limit_calls'] !== 0 && $files_token_file_data['calls'] >= $files_token_file_data['limit_calls'])) {
-                  return $response->withJson($api_sys_messages['AccessTokenInvalid'], 401);
-              }
+                // Delete folder
+                $delete_folder = flextype('media_folders')->delete($path);
 
-              if ($access_token_file_data['state'] === 'disabled' ||
-                  ($access_token_file_data['limit_calls'] !== 0 && $access_token_file_data['calls'] >= $access_token_file_data['limit_calls'])) {
-                  return $response->withJson($api_sys_messages['AccessTokenInvalid'], 401);
-              }
+                // Set response code
+                $response_code = $delete_folder ? 204 : 404;
 
-              // Delete folder
-              $delete_folder = $flextype['media_folders']->delete($id);
+                // Update calls counter
+                filesystem()->file($folders_token_file_path)->put(flextype('yaml')->encode(array_replace_recursive($folders_token_file_data, ['calls' => $folders_token_file_data['calls'] + 1])));
 
-              // Set response code
-              $response_code = ($delete_folder) ? 204 : 404;
+                if ($response_code === 404) {
+                    // Return response
+                    return $response
+                                ->withStatus($api_errors['0602']['http_status_code'])
+                                ->withHeader('Content-Type', 'application/json;charset=' . flextype('registry')->get('flextype.settings.charset'))
+                                ->write(flextype('json')->encode($api_errors['0602']));
+                }
 
-              // Update calls counter
-              Filesystem::write($files_token_file_path, $flextype['serializer']->encode(array_replace_recursive($files_token_file_data, ['calls' => $files_token_file_data['calls'] + 1]), 'yaml'));
+                // Return response
+                return $response
+                     ->withStatus($response_code)
+                ->withHeader('Content-Type', 'application/json;charset=' . flextype('registry')->get('flextype.settings.charset'))
+                ->write(flextype('json')->encode($delete_folder));
+            }
 
-              if ($response_code == 404) {
+            return $response
+                        ->withStatus($api_errors['0003']['http_status_code'])
+                        ->withHeader('Content-Type', 'application/json;charset=' . flextype('registry')->get('flextype.settings.charset'))
+                        ->write(flextype('json')->encode($api_errors['0003']));
+        }
 
-                  // Return response
-                  return $response
-                         ->withJson($api_sys_messages['NotFound'], $response_code);
-              }
+        return $response
+                    ->withStatus($api_errors['0003']['http_status_code'])
+                    ->withHeader('Content-Type', 'application/json;charset=' . flextype('registry')->get('flextype.settings.charset'))
+                    ->write(flextype('json')->encode($api_errors['0003']));
+    }
 
-              // Return response
-              return $response
-                     ->withJson($delete_file, $response_code);
-          }
-
-          return $response
-                 ->withJson($api_sys_messages['AccessTokenInvalid'], 401);
-      }
-
-      return $response
-             ->withJson($api_sys_messages['AccessTokenInvalid'], 401);
-  }
-
-  return $response
-         ->withJson($api_sys_messages['AccessTokenInvalid'], 401);
+    return $response
+                ->withStatus($api_errors['0003']['http_status_code'])
+                ->withHeader('Content-Type', 'application/json;charset=' . flextype('registry')->get('flextype.settings.charset'))
+                ->write(flextype('json')->encode($api_errors['0003']));
 });
