@@ -9,17 +9,19 @@ declare(strict_types=1);
 
 namespace Flextype;
 
-use Flextype\Component\Filesystem\Filesystem;
-use Psr\Http\Message\ResponseInterface as Response;
 use Psr\Http\Message\ServerRequestInterface as Request;
+use Slim\Http\Response;
+
 use function array_replace_recursive;
+use function filesystem;
+use function flextype;
 
 /**
  * Validate registry token
  */
-function validate_registry_token($token) : bool
+function validate_registry_token($token): bool
 {
-    return Filesystem::has(PATH['project'] . '/tokens/registry/' . $token . '/token.yaml');
+    return filesystem()->file(PATH['project'] . '/tokens/registry/' . $token . '/token.yaml')->exists();
 }
 
 /**
@@ -34,49 +36,58 @@ function validate_registry_token($token) : bool
  * Returns:
  * An array of registry item objects.
  */
-$app->get('/api/registry', function (Request $request, Response $response) use ($flextype, $api_sys_messages) {
-
+flextype()->get('/api/registry', function (Request $request, Response $response) use ($api_errors) {
     // Get Query Params
     $query = $request->getQueryParams();
+
+    if (! isset($query['id']) || ! isset($query['token'])) {
+        return $response->withStatus($api_errors['0300']['http_status_code'])
+                        ->withHeader('Content-Type', 'application/json;charset=' . flextype('registry')->get('flextype.settings.charset'))
+                        ->write(flextype('json')->encode($api_errors['0300']));
+    }
 
     // Set variables
     $id    = $query['id'];
     $token = $query['token'];
 
-    if ($flextype['registry']->get('flextype.settings.api.registry.enabled')) {
-
+    if (flextype('registry')->get('flextype.settings.api.registry.enabled')) {
         // Validate  token
         if (validate_registry_token($token)) {
             $registry_token_file_path = PATH['project'] . '/tokens/registry/' . $token . '/token.yaml';
 
             // Set  token file
-            if ($registry_token_file_data = $flextype['serializer']->decode(Filesystem::read($registry_token_file_path), 'yaml')) {
-                if ($registry_token_file_data['state'] === 'disabled' ||
-                    ($registry_token_file_data['limit_calls'] !== 0 && $registry_token_file_data['calls'] >= $registry_token_file_data['limit_calls'])) {
-                    return $response->withJson($api_sys_messages['AccessTokenInvalid'], 401);
+            if ($registry_token_file_data = flextype('yaml')->decode(filesystem()->file($registry_token_file_path)->get())) {
+                if (
+                    $registry_token_file_data['state'] === 'disabled' ||
+                    ($registry_token_file_data['limit_calls'] !== 0 && $registry_token_file_data['calls'] >= $registry_token_file_data['limit_calls'])
+                ) {
+                    return $response->withStatus($api_errors['0003']['http_status_code'])
+                                    ->withHeader('Content-Type', 'application/json;charset=' . flextype('registry')->get('flextype.settings.charset'))
+                                    ->write(flextype('json')->encode($api_errors['0003']));
                 }
 
                 // Fetch registry
-                if ($flextype['registry']->has($id)) {
+                if (flextype('registry')->has($id)) {
                     $response_data['data']['key']   = $id;
-                    $response_data['data']['value'] = $flextype['registry']->get($id);
+                    $response_data['data']['value'] = flextype('registry')->get($id);
 
                     // Set response code
                     $response_code = 200;
-
                 } else {
                     $response_data = [];
                     $response_code = 404;
                 }
 
                 // Update calls counter
-                Filesystem::write($registry_token_file_path, $flextype['serializer']->encode(array_replace_recursive($registry_token_file_data, ['calls' => $registry_token_file_data['calls'] + 1]), 'yaml'));
+                filesystem()->file($registry_token_file_path)
+                                      ->put(flextype('yaml')->encode(array_replace_recursive($registry_token_file_data, ['calls' => $registry_token_file_data['calls'] + 1])));
 
-                if ($response_code == 404) {
-
+                if ($response_code === 404) {
                     // Return response
                     return $response
-                           ->withJson($api_sys_messages['NotFound'], $response_code);
+                           ->withStatus($api_errors['0302']['http_status_code'])
+                           ->withHeader('Content-Type', 'application/json;charset=' . flextype('registry')->get('flextype.settings.charset'))
+                           ->write(flextype('json')->encode($api_errors['0302']));
                 }
 
                 // Return response
@@ -85,13 +96,19 @@ $app->get('/api/registry', function (Request $request, Response $response) use (
             }
 
             return $response
-                   ->withJson($api_sys_messages['AccessTokenInvalid'], 401);
+                   ->withStatus($api_errors['0003']['http_status_code'])
+                   ->withHeader('Content-Type', 'application/json;charset=' . flextype('registry')->get('flextype.settings.charset'))
+                   ->write(flextype('json')->encode($api_errors['0003']));
         }
 
         return $response
-               ->withJson($api_sys_messages['AccessTokenInvalid'], 401);
+               ->withStatus($api_errors['0003']['http_status_code'])
+               ->withHeader('Content-Type', 'application/json;charset=' . flextype('registry')->get('flextype.settings.charset'))
+               ->write(flextype('json')->encode($api_errors['0003']));
     }
 
     return $response
-           ->withJson($api_sys_messages['AccessTokenInvalid'], 401);
+            ->withStatus($api_errors['0003']['http_status_code'])
+            ->withHeader('Content-Type', 'application/json;charset=' . flextype('registry')->get('flextype.settings.charset'))
+            ->write(flextype('json')->encode($api_errors['0003']));
 });
